+++ app/components/ProjectAccessTab.tsx
@@ -21,6 +21,8 @@ import {
   useUserRows,
   type IdentityType,
   type Policy,
+  type RoleKey,
+  type RoleSource,
   type UserAccessRow,
 } from '@oxide/api'
 import { Badge } from '@oxide/design-system/ui'
@@ -37,7 +39,6 @@ import { confirmDelete } from '~/stores/confirm-delete'
 import { addToast } from '~/stores/toast'
 import { getActionsCol } from '~/table/columns/action-col'
 import { Table } from '~/table/Table'
-import type { IdentityFilter, ProjectAccessRow } from '~/types/access'
 import { CreateButton } from '~/ui/lib/CreateButton'
 import { TableActions } from '~/ui/lib/Table'
 import { TipIcon } from '~/ui/lib/TipIcon'
@@ -46,8 +47,17 @@ import {
   identityFilterLabel,
   identityTypeLabel,
   roleColor,
+  type IdentityFilter,
 } from '~/util/access'
 import { groupBy } from '~/util/array'
+
+type ProjectAccessRow = {
+  id: string
+  identityType: IdentityType
+  name: string
+  projectRole: RoleKey | undefined
+  roleBadges: { roleSource: RoleSource; roleName: RoleKey }[]
+}
 
 type ProjectAccessTabProps = {
   filter: IdentityFilter
